Bug fix
Im going to test on my laptop as well to be sure by sending an receiving this commit. However the duplicate objects issue is not present on my main machine anymore. The solve was ensuring that the arrayLists are cleared on entry of code that will be building the lists anyway.